refactor(UploadQueue): Extract code from main component
Optimize Item ListItem code too
